extend librtlsdr API: set tuner stage gains
add rtlsdr_get_tuner_stage_gains() to get a description of the
internal hardware tuner stages and their gain settings.

rtlsdr_set_tuner_stage_gain() will set the gain of a certain stage.

TODO: dev->gain is not yet updated

Always fill in the stage description if a buffer was provided, even when not returning gains.